try converting abilitybot
